Refactor game view
