+++ src/components/Game/GameView.tsx
@@ -1,12 +1,19 @@
 import { makeStyles } from '@material-ui/core';
-import React, { useContext, useEffect, useMemo, useRef, useState } from 'react';
+import React, {
+  useCallback,
+  useContext,
+  useEffect,
+  useMemo,
+  useRef,
+  useState,
+} from 'react';
 import * as Tone from 'tone';
 import { Player } from 'tone';
 import { noOp } from 'tone/build/esm/core/util/Interface';
 import { PlayerContext } from '../../contexts/PlayerContext';
 import { TraditionalKeyboardDimension } from '../../types/keyboardDimension';
 import { Part } from '../../types/messages';
-import { MidiJSON, Note } from '../../types/MidiJSON';
+import { MidiJSON } from '../../types/MidiJSON';
 import { calculateSmartKeyboardDimension } from '../../utils/calculateSmartKeyboardDimension';
 import { calculateTraditionalKeyboardDimensionForGame } from '../../utils/calculateTraditionalKeyboardDimension';
 import { isEqual } from '../../utils/setHelpers';
@@ -14,14 +21,8 @@ import { calculateSongDuration, getPlaybackNotes } from '../../utils/songInfo';
 import { useDimensions } from '../../utils/useDimensions';
 import InstrumentPlayer from '../Piano/InstrumentPlayer';
 import { NullSoundFontPlayerNoteAudio } from '../Piano/InstrumentPlayer/AudioPlayer';
-import { MappedNote } from '../Piano/types/mappedNote';
-import { getIndexToNotesMap } from '../Piano/utils/getKeyToNotesMap';
 import { changeSongSpeed } from '../utils';
-import { IndexedNote } from '../Waterfall/types';
-import {
-  calculateLookAheadTime,
-  getIndexedNotesFromNotes,
-} from '../Waterfall/utils';
+import { calculateLookAheadTime } from '../Waterfall/utils';
 import GameMiddleView from './GameMiddleView';
 import GameSmartPiano from './GameSmartPiano';
 import GameTraditionalPiano from './GameTraditionalPiano';
@@ -68,12 +69,17 @@ const GameView: React.FC<Props> = ({
 }) => {
   const classes = useStyles();
 
+  // Game start time (after the countdown)
   const [startTime, setStartTime] = useState(-1);
   const [gameEnd, setGameEnd] = useState(false);
   const countDown = 3;
   const [timeToStart, setTimeToStart] = useState(countDown);
   const pressedNotes = useRef<Set<number>>(new Set());
-  const modifiedMIDI = changeSongSpeed(chosenSongMIDI, speed);
+  const modifiedMIDI = useMemo(
+    () => changeSongSpeed(chosenSongMIDI, speed),
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    []
+  );
 
   // For scoring
   const gameEndRef = useRef(false);
@@ -91,6 +97,7 @@ const GameView: React.FC<Props> = ({
     return calculateLookAheadTime(header);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
+  // Actual start time (including the lookAheadTime(1 bar rest))
   const delayedStartTime = lookAheadTime + startTime;
 
   /*************** Player track information *****************/
@@ -100,25 +107,11 @@ const GameView: React.FC<Props> = ({
   // 1 for solo, 2 for secondo
   const playbackChannel = myPart === undefined ? 1 : 2;
   const playerTrack = tracks[playerTrackNum];
-  const normalPlayerNotes = playerTrack.notes;
-  const indexToNotesMap: MappedNote[][] = useMemo(
-    () => getIndexToNotesMap(normalPlayerNotes),
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    []
-  );
-
-  const playerNotes = useMemo<Note[] | IndexedNote[]>(() => {
-    if (showSmartPiano) {
-      return getIndexedNotesFromNotes(normalPlayerNotes);
-    } else {
-      return normalPlayerNotes;
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  const playerNotes = playerTrack.notes;
 
   /*************** Initialise instrument *****************/
   // TODO: schedule change (if have time), now take the first value only
-  const keyboardVolume = normalPlayerNotes[0].velocity;
+  const keyboardVolume = playerNotes[0].velocity;
   const instrumentPlayer = useMemo(
     () => new InstrumentPlayer(keyboardVolume),
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -166,11 +159,6 @@ const GameView: React.FC<Props> = ({
       }, note.time + delayedStartTime - Tone.now() - 1);
     });
 
-    // Schedule ending screen
-    Tone.Transport.schedule(() => {
-      setGameEnd(true);
-    }, delayedStartTime + songDuration - Tone.now() + 0.1);
-
     // Score update
     const scoreHandler = setInterval(() => {
       // stop updating score if game has ended
@@ -190,7 +178,7 @@ const GameView: React.FC<Props> = ({
       // we use refs here to reduce computation workload during each callback
       const [correctNotes, index] = getNotesAtTimeFromNotes(
         Tone.now() - delayedStartTime,
-        normalPlayerNotes,
+        playerNotes,
         prevIndexInMIDI.current
       );
 
@@ -217,30 +205,36 @@ const GameView: React.FC<Props> = ({
 
   /*************** Callbacks *****************/
 
-  const didPlayNote = (note: number, playedBy: number) => {
-    if (playedBy === me) {
-      pressedNotes.current.add(note);
-    }
-    handleNotePlay(note, playedBy);
-  };
+  const didPlayNote = useCallback(
+    (note: number, playedBy: number) => {
+      if (playedBy === me) {
+        pressedNotes.current.add(note);
+      }
+      handleNotePlay(note, playedBy);
+    },
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [me]
+  );
 
-  const didStopNote = (note: number, playedBy: number) => {
-    if (playedBy === me) {
-      pressedNotes.current.delete(note);
-    }
-    handleNoteStop(note, playedBy);
-  };
+  const didStopNote = useCallback(
+    (note: number, playedBy: number) => {
+      if (playedBy === me) {
+        pressedNotes.current.delete(note);
+      }
+      handleNoteStop(note, playedBy);
+    },
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [me]
+  );
 
   /*************** Keyboard dimension *****************/
   const [middleBoxDimensions, middleBoxRef] = useDimensions<HTMLDivElement>();
   const keyboardDimension = useMemo(() => {
+    const { width } = middleBoxDimensions;
     if (showSmartPiano) {
-      return calculateSmartKeyboardDimension(middleBoxDimensions.width);
+      return calculateSmartKeyboardDimension(width);
     } else {
-      return calculateTraditionalKeyboardDimensionForGame(
-        middleBoxDimensions.width,
-        playerTrack
-      );
+      return calculateTraditionalKeyboardDimensionForGame(width, playerTrack);
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [middleBoxDimensions.width]);
@@ -252,7 +246,7 @@ const GameView: React.FC<Props> = ({
         <GameSmartPiano
           instrumentPlayer={instrumentPlayer}
           keyWidth={keyboardDimension.keyWidth}
-          indexToNotesMap={indexToNotesMap}
+          normalPlayerNotes={playerNotes}
           didPlayNote={didPlayNote}
           didStopNote={didStopNote}
           startTime={delayedStartTime}
@@ -291,7 +285,7 @@ const GameView: React.FC<Props> = ({
           startTime={startTime}
           lookAheadTime={lookAheadTime}
           keyboardDimension={keyboardDimension}
-          playerNotes={playerNotes}
+          normalPlayerNotes={playerNotes}
         />
       </div>
       <div className={classes.piano}>{!gameEnd && piano}</div>
